+++ R/ukhousing-package.R
@@ -2,6 +2,7 @@
 "_PACKAGE"
 
 #' @importFrom cli cli_abort cli_inform cli_warn
+#' @importFrom jsonlite fromJSON
 #' @importFrom stats median setNames
 #' @importFrom tools R_user_dir
 #' @importFrom utils read.csv unzip URLencode